lots of new assets
the assets will not be staying in this configuration, they are simply placeholders that will eventually be the normal tiles. we need to use them as placeholders to determine the best route for placing them in the game whether it be via tile sets or as single objects. As for the actual tiles, they should be kept as tile sets, and objects like boxes trees and rocks will be placed as individual objects that will be more extensible with code.